move proto.util.render to Schema
